Tests of local appium DriverService were re-designed.
