--- src/test/java/io/appium/java_client/localserver/custom_node_path.properties
@@ -1,19 +1,0 @@
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# See the NOTICE file distributed with this work for additional
-# information regarding copyright ownership.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-path.to.custom.node.win=C:/Program Files (x86)/Appium/node_modules/appium/bin/appium.js
-path.to.custom.node.macos=/Applications/Appium.app/Contents/Resources/node_modules/appium/bin/appium.js
-path.to.custom.node.linux=specify your path on your own
